Revert "Testem e caso não dê erro pode aceitar"
