+++ Codigo/GestaoAluguel/GestaoAluguelWeb/Views/Home/Index.cshtml
@@ -7,9 +7,4 @@
     <p>Learn about <a href="https://learn.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
 
     <a asp-controller="Pessoa" asp-action="Create">Cadastre-se!</a>
-    <a asp-controller="Imovel" asp-action="Index"> Ver Imoveis</a>
-    <a asp-controller="Locacao" asp-action="Index"> Ver Locacoes</a>
-    <p>Ola mundo</p>
-
-
 </div>
