+++ Codigo/GestaoAluguel/GestaoAluguelWeb/Views/Locacao/Details.cshtml
@@ -67,6 +67,6 @@
     </dl>
 </div>
 <div>
-    @Html.ActionLink("Edit", "Edit", new { /* id = Model.PrimaryKey */ }) |
+    @Html.ActionLink("Edit", "Edit", new { id = Model.Id }) |
     <a asp-action="Index">Back to List</a>
 </div>
