--- Codigo/GestaoAluguel/GestaoAluguelWeb/Views/Locacao/FinalizarLocacao.cshtml
@@ -1,105 +1,0 @@
-﻿@* @using GestaoAluguelWeb.Helpers
-@model GestaoAluguelWeb.Models.LocacaoModel
-@{
-    ViewData["Title"] = "Finalizar Locacao";
-}
-
-<div class="container mt-4">
-    <h2>@ViewData["Title"]</h2>
-    <hr />
-
-    @if (Model.Locacao != null)
-    {
-        <div class="row">
-            <div class="col-md-6">
-                <div class="card mb-4">
-                    <div class="card-header bg-primary text-white">
-                        <h5>Dados da Locação</h5>
-                    </div>
-                    <div class="card-body">
-                        <dl class="row">
-                            <dt class="col-sm-4">Código:</dt>
-                            <dd class="col-sm-8">@Model.LocacaoModel.Id</dd>
-
-                            <dt class="col-sm-4">Imóvel:</dt>
-                            <dd class="col-sm-8">@Model.imovelInfo</dd>
-
-                            <dt class="col-sm-4">Inquilino:</dt>
-                            <dd class="col-sm-8">@Model.InquilinoInfo</dd>
-
-                            <dt class="col-sm-4">Data Início:</dt>
-                            <dd class="col-sm-8">@Model.Locacao.DataInicio.ToString("dd/MM/yyyy")</dd>
-
-                            <dt class="col-sm-4">Valor do Aluguel:</dt>
-                            <dd class="col-sm-8">R$ @Model.Locacao.ValorAluguel.ToString("N2")</dd>
-
-                            <dt class="col-sm-4">Status:</dt>
-                            <dd class="col-sm-8">@Model.StatusLocacao</dd>
-                        </dl>
-                    </div>
-                </div>
-            </div>
-
-            <div class="col-md-6">
-                <div class="card mb-4">
-                    <div class="card-header bg-success text-white">
-                        <h5>Finalizar Locação</h5>
-                    </div>
-                    <div class="card-body">
-                        <form method="post" enctype="multipart/form-data">
-                            <input type="hidden" asp-for="Id" value="@Model.Locacao.Id" />
-
-                            <div class="form-group">
-                                <label asp-for="DataFim" class="control-label">Data de Finalização</label>
-                                <input asp-for="DataFim" class="form-control" type="date"
-                                       value="@DateTime.Now.ToString("yyyy-MM-dd")" />
-                                <span asp-validation-for="DataFim" class="text-danger"></span>
-                            </div>
-
-                            <div class="form-group mt-3">
-                                <label asp-for="Motivo" class="control-label">Motivo da Finalização</label>
-                                <textarea asp-for="Motivo" class="form-control" rows="3"
-                                          placeholder="Informe o motivo da finalização da locação"></textarea>
-                                <span asp-validation-for="Motivo" class="text-danger"></span>
-                            </div>
-
-                            <div class="form-group mt-3">
-                                <label class="control-label">Contrato de Encerramento (opcional)</label>
-                                <input type="file" asp-for="ContratoEncerramento" class="form-control"
-                                       accept=".pdf,.doc,.docx,.jpg,.png" />
-                                <small class="form-text text-muted">Formatos aceitos: PDF, DOC, DOCX, JPG, PNG</small>
-                            </div>
-
-                            <div class="alert alert-info mt-3">
-                                <h6>Resumo:</h6>
-                                <p><strong>Duração da Locação:</strong> @Model.DuracaoLocacao dias</p>
-                                <p><strong>Valor Total Estimado:</strong> R$ @Model.ValorTotalEstimado.ToString("N2")</p>
-                            </div>
-
-                            <div class="form-group mt-3">
-                                <button type="submit" class="btn btn-success btn-lg">
-                                    <i class="fas fa-check-circle"></i> Finalizar Locação
-                                </button>
-                                <a asp-page="/Locacoes/Index" class="btn btn-secondary">Cancelar</a>
-                            </div>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        </div>
-    }
-    else
-    {
-        <div class="alert alert-warning">
-            <h4>Locação não encontrada!</h4>
-            <p>A locação solicitada não foi encontrada ou já foi finalizada.</p>
-            <a asp-page="/Locacoes/Index" class="btn btn-primary">Voltar para Lista</a>
-        </div>
-    }
-</div>
-
-@section Scripts {
-    @{
-        await Html.RenderPartialAsync("_ValidationScriptsPartial");
-    }
-} *@
