+++ Codigo/GestaoAluguel/GestaoAluguelWeb/Views/Locacao/Index.cshtml
@@ -7,7 +7,8 @@
 <h1>Index</h1>
 
 <p>
-    <a asp-action="Create">Create New</a>
+    <a asp-action="Create" class="btn btn-primary">Create New</a>
+    <a asp-controller="Imovel" asp-action="Index" class="btn btn-secondary ms-2">Ver Imóveis</a>
 </p>
 <table class="table">
     <thead>
@@ -73,9 +74,9 @@
                 @Html.DisplayFor(modelItem => item.Contrato)
             </td>
             <td>
-                @Html.ActionLink("Edit", "Edit", new { /* id=item.PrimaryKey */ }) |
-                @Html.ActionLink("Details", "Details", new { /* id=item.PrimaryKey */ }) |
-                @Html.ActionLink("Delete", "Delete", new { /* id=item.PrimaryKey */ })
+                    @Html.ActionLink("Edit", "Edit", new { id = item.Id }) |
+                    @Html.ActionLink("Details", "Details", new { id = item.Id }) |
+                    @*Html.ActionLink("Delete", "Delete", new { id = item.Id })*@
             </td>
         </tr>
 }
