--- Codigo/GestaoAluguel/GestaoAluguelWeb/Views/Locacao2/Create.cshtml
@@ -1,65 +1,0 @@
-﻿@model GestaoAluguelWeb.Models.LocacaoModel
-
-@{
-    ViewData["Title"] = "Create";
-}
-
-<h1>Create</h1>
-
-<h4>LocacaoModel</h4>
-<hr />
-<div class="row">
-    <div class="col-md-4">
-        <form asp-action="Create">
-            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            
-            @* Campos ocultos *@
-            <input type="hidden" asp-for="Id" />
-            <input type="hidden" asp-for="IdImovel" />
-            <input type="hidden" asp-for="Status" />
-
-            <div class="form-group">
-                <label asp-for="DataInicio" class="control-label"></label>
-                <input asp-for="DataInicio" class="form-control" />
-                <span asp-validation-for="DataInicio" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="DataFim" class="control-label"></label>
-                <input asp-for="DataFim" class="form-control" />
-                <span asp-validation-for="DataFim" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="ValorAluguel" class="control-label"></label>
-                <input asp-for="ValorAluguel" class="form-control" />
-                <span asp-validation-for="ValorAluguel" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="IdInquilino" class="control-label"></label>
-                <input asp-for="IdInquilino" class="form-control" />
-                <span asp-validation-for="IdInquilino" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Motivo" class="control-label"></label>
-                <input asp-for="Motivo" class="form-control" />
-                <span asp-validation-for="Motivo" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Contrato" class="control-label"></label>
-                <input asp-for="Contrato" class="form-control" />
-                <span asp-validation-for="Contrato" class="text-danger"></span>
-            </div>
-            <div class="form-group d-flex gap-2 mt-2">
-                <input type="submit" value="Alugar Imóvel" class="btn btn-success" />
-                <a class="btn btn-secondary"
-                   asp-controller="Imovel"
-                   asp-action="Index">
-                    Voltar
-                </a>
-            </div>
-        </form>
-    </div>
-</div>
-
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
